rename custom vuexfire to vuexfireJAC
